displaying imgs for store and changes to image retriever

+++ src/java/security/filter/resource/ImageRetrieverFilter.java
@@ -2,6 +2,8 @@ package security.filter.resource;
 
 import controller.redirect.ErrorRedirect;
 import controller.resource.Resources;
+
+import java.io.File;
 import java.io.IOException;
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
@@ -12,11 +14,13 @@ import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import obj.account.Account;
 import security.error.Errors;
+import util.CheckFormat;
+import util.UserInput;
 
 public class ImageRetrieverFilter implements Filter {
 
-    private final String IMAGE_PATH_AVA = "users\\ava\\";
-    private final String IMAGE_EXT = ".jpg";
+    private final String IMAGE_PATH_AVA = "users\\";
+    private final String IMAGE_PATH_PLANT = "plants\\";
 
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
@@ -31,7 +35,6 @@ public class ImageRetrieverFilter implements Filter {
             return;
         }
 
-        String imageName;
         StringBuilder builder = new StringBuilder();
         builder.append(DBPath);
 
@@ -42,23 +45,46 @@ public class ImageRetrieverFilter implements Filter {
 
                 if (account == null) {
                     destroyLocalDBPath(request);
-                    ErrorRedirect.redirect(Errors.ACCESS_DENIED, request, response);
+                    ErrorRedirect.redirect(Errors.BAD_REQUEST, request, response);
                     return;
                 }
 
-                imageName = String.valueOf(account.getID());
                 builder.append(IMAGE_PATH_AVA);
+                builder.append(Integer.toString(account.getID()));
+                break;
+
+            case PLANT:
+                String pid = request.getParameter("pid");
+
+                if (UserInput.isEmpty(pid)) {
+                    destroyLocalDBPath(request);
+                    ErrorRedirect.redirect(Errors.BAD_REQUEST, request, response);
+                    return;
+                }
+
+                builder.append(IMAGE_PATH_PLANT);
+                builder.append(pid);
                 break;
 
             default:
-                destroyLocalDBPath(request);
-                ErrorRedirect.redirect(Errors.BAD_REQUEST, request, response);
-                return;
+                throw new ServletException();
         }
-        
-        builder.append(imageName);
-        builder.append(IMAGE_EXT);
-        request.setAttribute("imagePath", builder.toString());
+
+        File folder = new File(builder.toString());
+
+        if (!folder.isDirectory()) {
+            destroyLocalDBPath(request);
+            ErrorRedirect.redirect(Errors.FILE_NOT_FOUND, request, response);
+            return;
+        }
+
+        File[] files = folder.listFiles();
+
+        for (File file : files) {
+            if (CheckFormat.check(file.getName(), "^img[.]")) {
+                request.setAttribute("imgFile", file);
+            }
+        }
 
         destroyLocalDBPath(request);
         chain.doFilter(request, response);
